requirements.txt
streamlit
Pillow
setuptools
wheel